fix bug: không mở được showshow

+++ ZaloPay/app/src/main/java/vn/com/vng/zalopay/paymentapps/PaymentAppConfig.java
@@ -61,7 +61,7 @@ public class PaymentAppConfig {
     }
 
     public static AppResource getAppResource(long appId) {
-        return APP_RESOURCE_MAP.get(appId);
+        return APP_RESOURCE_MAP.get((int) appId);
     }
 
 }
